Add a driver interface. Fix a possible bug in the elf loader. Make a hash table (just to find out I don't need it). Make a ps/2 driver. Update build instructions.

+++ src/oboskrnl/descriptors/idt/pic.cpp
@@ -137,7 +137,7 @@ namespace obos
 		return utils::testBitInBitfield(isr_register, interrupt);
 	}
 
-	void SendEOI(DWORD irqNumber)
+	void SendEOI(BYTE irqNumber)
 	{
 		if (irqNumber > 15)
 			return;
